+++ src/common/services/WorkItemService.ts
@@ -1,3 +1,4 @@
+import { DevOpsService, IDevOpsService } from '@joachimdalen/azdevops-ext-core/DevOpsService';
 import { getClient } from 'azure-devops-extension-api/Common';
 import { CoreRestClient, ProjectProperty } from 'azure-devops-extension-api/Core';
 import { IdentityRef } from 'azure-devops-extension-api/WebApi';
@@ -11,7 +12,6 @@ import {
 import { WorkItemTrackingProcessRestClient } from 'azure-devops-extension-api/WorkItemTrackingProcess';
 
 import { getChildIds, getParentId } from '../workItemUtils';
-import DevOpsService, { IDevOpsService } from './DevOpsService';
 
 export interface IWorkItemService {
   getParentForWorkItem(id: number, workItem?: WorkItem): Promise<WorkItem | undefined>;